Aplicando algunos ajustes después de incluir el nuevo componente de ScheduledProblemList
